requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
huggingface_hub
pdfplumber
python-docx